@@ -2,7 +2,8 @@ package com.jpmcosta.test.realmtestproject;
 
 import android.app.Application;
 import android.content.Intent;
-import android.support.annotation.NonNull;
+
+import androidx.annotation.NonNull;
 
 import com.jpmcosta.test.realmtestproject.realm.App;
 import com.jpmcosta.test.realmtestproject.realm.Item;
