@@ -1,9 +1,10 @@
 package com.jpmcosta.test.realmtestproject.activity;
 
 import android.os.Bundle;
-import android.support.v7.widget.LinearLayoutManager;
-import android.support.v7.widget.RecyclerView;
 import android.widget.TextView;
+
+import androidx.recyclerview.widget.LinearLayoutManager;
+import androidx.recyclerview.widget.RecyclerView;
 
 import com.jpmcosta.test.realmtestproject.R;
 import com.jpmcosta.test.realmtestproject.adapter.ItemListAdapter;
