@@ -1,11 +1,12 @@
 package com.jpmcosta.test.realmtestproject.adapter;
 
-import android.support.annotation.NonNull;
-import android.support.v7.widget.RecyclerView;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.TextView;
+
+import androidx.annotation.NonNull;
+import androidx.recyclerview.widget.RecyclerView;
 
 import com.jpmcosta.test.realmtestproject.R;
 import com.jpmcosta.test.realmtestproject.realm.Item;
